Create README.md
VBA script that analyzes certain tickers for the years 2014, 2015, and 2016 and outputs the yearly change, percent change, and total stock volume for each ticker during these years. The  VBA script will also output the the ticker that had the  greatest % increase, greatest % decrease and greatest total volume for each of these years.